+++ packages/plugin-data/CHANGELOG.md
@@ -3,15 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 0.2.0 (2020-11-24)
+## [0.2.5](https://github.com/salesforcecli/data/compare/@salesforce/plugin-data@0.2.4...@salesforce/plugin-data@0.2.5) (2020-12-15)
 
+**Note:** Version bump only for package @salesforce/plugin-data
+
+## [0.2.4](https://github.com/salesforcecli/data/compare/@salesforce/plugin-data@0.2.3...@salesforce/plugin-data@0.2.4) (2020-12-15)
+
+**Note:** Version bump only for package @salesforce/plugin-data
+
+## [0.2.3](https://github.com/salesforcecli/data/compare/@salesforce/plugin-data@0.2.2...@salesforce/plugin-data@0.2.3) (2020-12-15)
+
+**Note:** Version bump only for package @salesforce/plugin-data
+
+## [0.2.2](https://github.com/salesforcecli/data/compare/@salesforce/plugin-data@0.2.1...@salesforce/plugin-data@0.2.2) (2020-12-10)
+
+**Note:** Version bump only for package @salesforce/plugin-data
+
+## [0.2.1](https://github.com/salesforcecli/data/compare/@salesforce/plugin-data@0.2.0...@salesforce/plugin-data@0.2.1) (2020-12-03)
+
+### Bug Fixes
+
+- various data:record bugs ([a05ac07](https://github.com/salesforcecli/data/commit/a05ac07e7beb9d73776908beed085b7644d580fe))
+
+# 0.2.0 (2020-11-24)
 
 ### Features
 
-* record commands ([37f5d61](https://github.com/salesforcecli/data/commit/37f5d61f65055c76aceac85077fd484c810bd49c))
-
-
-
-
+- record commands ([37f5d61](https://github.com/salesforcecli/data/commit/37f5d61f65055c76aceac85077fd484c810bd49c))
 
 # Change Log